rag-api/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
openai
faiss-cpu
tiktoken
python-dotenv
python-multipart